+++ CHANGELOG.md
@@ -24,6 +24,15 @@ DO NOT TOUCH. SAVE IT ON TOP.
 
 -->
 
+## [3.0.0] - 2020-06-23
+### Changed
+- New exports signature;
+- New query parameters;
+- Some options were changed.
+
+### Added
+- Runtime helpers.
+
 ## [2.0.0] - 2020-05-26
 ### Breaking
 - Requires `Node 10+`